refactor(cyclomatic): homogenize file names
